Inject connection directly rather than by name.
The only reason for injecting the connection by name was to avoid issues
with marshalling. To me, it is not currently clear whether we actually
want to allow marshalling in this case. We could probably do something
like allow marshalling it if the connection was established from a
specification name, but for now we can just be conservative.