Don't attempt to manipulate the class level connection
There were several methods that `Relation` had overridden to set the
class level connection on its AR subclass before calling. The majority
of these methods are ones where we either control the implementation
(and don't expect it to be overridden)

The one case where we don't always control the implementation was
`prefetch_primary_key` and `next_sequence_value`. Interestingly, the
only consumer of those that I'm aware of (Oracle adapter) is using
`@connection` not the class method, so it wouldn't have picked up
changes from `use_connection` anyway.

I've intentionally left the schema related queries on the main
connection for now. We can assume that these methods will return the
same results regardless of the connection that it's run on, and I don't
think we're yet ready to answer the question of which connection these
will run on yet, and the answer to that question is going to depend on
which public APIs we end up introducing on top of that.